Completed game of war
Functionally complete, but the code is a mess which I will now attempt
to refactor.

I have stepped through the test case where each player starts with 3
cards, and it seems to work correctly.

However, I'm a bit surprised to see that when trying a few games with
random cards, the algorithm seems to very easily enter a loop and never
terminates. I'll investigate Carol's solution when I'm happy with my
own.